Deleted getPaymentMethods() from SDK class; Added getAll() to PaymentMethod resource;
