+++ src/main/java/com/mercadopago/resources/PaymentMethod.java
@@ -1,8 +1,21 @@
 package com.mercadopago.resources;
 
+import static com.mercadopago.MercadoPago.SDK.getBaseUrl;
+
+import com.google.gson.FieldNamingPolicy;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.reflect.TypeToken;
+import com.mercadopago.core.MPApiResponse;
+import com.mercadopago.core.annotations.rest.PayloadType;
+import com.mercadopago.exceptions.MPRestException;
+import com.mercadopago.net.HttpMethod;
+import com.mercadopago.net.MPRestClient;
 import com.mercadopago.resources.datastructures.paymentmethod.FinancialInstitutions;
 import com.mercadopago.resources.datastructures.paymentmethod.Settings;
+import java.lang.reflect.Type;
 import java.util.List;
+import org.apache.commons.lang3.StringUtils;
 
 public class PaymentMethod {
 
@@ -95,5 +108,30 @@ public class PaymentMethod {
 
   public List<String> getProcessingModes() {
     return processingModes;
+  }
+
+  /**
+   *
+   * @return List of payment methods
+   * @throws MPRestException exception
+   */
+  public static List<PaymentMethod> getAll() throws MPRestException {
+    String uri = StringUtils.join(getBaseUrl(), "/v1/payment_methods");
+    MPRestClient client = new MPRestClient();
+    MPApiResponse mpApiResponse = client.executeRequest(HttpMethod.GET, uri, PayloadType.JSON, null);
+    return parsePaymentMethodJson(mpApiResponse);
+  }
+
+  /**
+   *
+   * @param mpApiResponse response from api
+   * @return list of payment methods formatted
+   */
+  private static List<PaymentMethod> parsePaymentMethodJson(MPApiResponse mpApiResponse) {
+    Gson gson =  new GsonBuilder()
+        .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
+        .create();
+    Type listType = new TypeToken<List<PaymentMethod>>(){}.getType();
+    return gson.fromJson(mpApiResponse.getJsonElementResponse(), listType);
   }
 }
