+++ src/test/java/mercadopago/resources/PaymentMethodTest.java
@@ -3,7 +3,6 @@ package mercadopago.resources;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-import com.mercadopago.MercadoPago;
 import com.mercadopago.exceptions.MPRestException;
 import com.mercadopago.resources.PaymentMethod;
 import java.util.List;
@@ -13,7 +12,7 @@ public class PaymentMethodTest extends BaseResourceTest {
 
   @Test
   public void getPaymentMethods() throws MPRestException {
-    List<PaymentMethod> response = MercadoPago.SDK.getPaymentMethods();
+    List<PaymentMethod> response = PaymentMethod.getAll();
     assertNotNull(response);
     assertTrue(response.size() > 0);
   }
